break out GroupSerializer
