+++ api/serializers/__init__.py
@@ -30,3 +30,4 @@ from .atmo_user_serializer import AtmoUserSerializer
 from .profile_serializer import ProfileSerializer
 from .provider_machine_serializer import ProviderMachineSerializer
 from .paginated_provider_machine_serializer import PaginatedProviderMachineSerializer
+from .group_serializer import GroupSerializer
